requirements.txt
# Options Buddy - Dependencies

# Web Framework
streamlit>=1.30.0

# Interactive Brokers API
ib_insync>=0.9.86

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Options Pricing (Black-Scholes)
py_vollib>=1.0.1

# Database
sqlalchemy>=2.0.0

# Visualization
plotly>=5.18.0

# Environment Variables
python-dotenv>=1.0.0

# Date/Time Utilities
python-dateutil>=2.8.2

# Testing
pytest>=7.4.0
